{"schema":"decodex/commit/1","summary":"Modularize toolbar annotation style controls","authority":"manual"}
